+++ ClientExample/Unity3D/Components/Souers/TmTeacherController.cs
@@ -14,10 +14,6 @@ namespace ClientExample
                     Console.WriteLine(TmTimerTool.CurrentTime() + " TmTeacherController: " + elevenCode);
                     SetSoulerDBs(parameter);
                     break;
-                //case (ElevenCode.GetRolers):
-                //    Console.WriteLine(TmTimerTool.CurrentTime() + " TmTeacherController: " + elevenCode);
-                //    GetTeachers(parameter);
-                //    break;
                 default:
                     break;
             }
@@ -31,11 +27,6 @@ namespace ClientExample
                 Console.WriteLine(TmTimerTool.CurrentTime() + " TmObjects.Teachers: " + TmObjects.Teachers.Count);
             }
         }
-        //void GetTeachers(TmParameter parameter)
-        //{
-        //    List<TmSoulerDB> teachers = TmParameterTool.GetJsonValue<List<TmSoulerDB>>(parameter, parameter.ElevenCode.ToString());
-        //    Console.WriteLine(TmTimerTool.CurrentTime() + " TmTeacherController-Teachers: " + teachers.Count);
-        //}   
         
     }
 }
