--- ClientExample/Unity3D/Players/TmEngineer.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using Tumo;
-namespace ClientExample
-{
-    public class TmEngineer : TmEntity
-    {
-    }
-}
