+++ Servers/Games/Components/TmHandlers/Soulers/TmBookerHandler.cs
@@ -10,10 +10,6 @@ namespace Servers
             ElevenCode elevenCode = parameter.ElevenCode;
             switch (elevenCode)
             {
-                //case (ElevenCode.GetRolers):
-                //    Console.WriteLine(TmTimerTool.CurrentTime() + " TmBookerHandler: " + elevenCode);
-                //    //GetRolersByRolerId(parameter);
-                //    break;
                 case (ElevenCode.StatusSync):
                     Console.WriteLine(TmTimerTool.CurrentTime() + " TmStatusSyncHandler: " + elevenCode);
                     Parent.GetComponent<TmStatusSyncHandler>().OnTransferParameter(this, parameter);
@@ -25,7 +21,6 @@ namespace Servers
             }
         }
         internal Dictionary<int, TmSoulerDB> Bookers { get; set; }
-        public Dictionary<string, TmMonster> SpawnCDs = new Dictionary<string, TmMonster>();
         void GetRolersByRolerId(TmParameter parameter)
         {
             bool yes = false;
@@ -34,7 +29,6 @@ namespace Servers
             {
                 if (this.Bookers != null)
                 {
-                    //TmObjects.Bookers = this.Bookers;
                     TmParameter response = TmParameterTool.ToJsonParameter(TenCode.Booker, ElevenCode.GetRolers, ElevenCode.GetRolers.ToString(), this.Bookers);
                     response.Keys.Add(parameter.Keys[0]);
                     TmTcpSocket.Instance.Send(response);
@@ -57,10 +51,6 @@ namespace Servers
         void DiethHandler(TmParameter parameter)
         {
             TmSoulerDB soulerDB = TmParameterTool.GetJsonValue<TmSoulerDB>(parameter, ElevenCode.Die.ToString());
-            TmMonster monster = new TmMonster(this);
-            monster.AddComponent(soulerDB);
-            //(monster.GetComponent<TmCoolDown>() as TmCoolDown).Key = parameter.Key;
-            SpawnCDs.Add(monster.EcsId, monster);
 
             parameter.ElevenCode = ElevenCode.RolerRemove;
             foreach(var tem in TmTcpSocket.Instance.TPeers.Keys)
