+++ Servers/Games/Components/TmHandlers/TmMysqlHandler.cs
@@ -19,7 +19,6 @@ namespace Servers
             this.AddComponent(new TmEngineerMysql());
             this.AddComponent(new TmBookerMysql());
             this.AddComponent(new TmTeacherMysql());
-            this.AddComponent(new TmDressedMysql());
             this.AddComponent(new TmKnapsackMysql());
             this.AddComponent(new TmSmithyMysql());
             this.AddComponent(new TmAbilityMysql());
