+++ Servers/Mysqls/Components/GridMaps/TmGridMysql.cs
@@ -11,7 +11,6 @@ namespace Servers
     {
         public override void TmAwake()
         {
-            base.TmAwake();
             DatabaseFormName = "gridmap";
         }
         public override void OnTransferParameter(object sender, TmParameter parameter)
@@ -19,17 +18,9 @@ namespace Servers
             ElevenCode elevenCode = parameter.ElevenCode;
             switch (elevenCode)
             {
-                case (ElevenCode.Get):
+                case (ElevenCode.Map):
                     Console.WriteLine(TmTimerTool.CurrentTime() + " TmEngineerMysql: " + elevenCode);
-                    //GetSouler(sender, parameter);
-                    break;
-                case (ElevenCode.GetRolers):
-                    Console.WriteLine(TmTimerTool.CurrentTime() + " TmEngineerMysql: " + elevenCode);
-                    GetRolersByUersId(sender, parameter);
-                    break;
-                case (ElevenCode.EngineerLogin):
-                    Console.WriteLine(TmTimerTool.CurrentTime() + " TmEngineerMysql: " + elevenCode);
-                    //GetdbofEngineerLogin(sender, parameter);
+                    GetGridMap(this, parameter);
                     break;
                 case (ElevenCode.None):
                     break;
@@ -37,7 +28,7 @@ namespace Servers
                     break;
             }
         }
-        void GetRolersByUersId(object sender, TmParameter parameter)
+        void GetGridMap(object sender, TmParameter parameter)
         {
             int userId = TmParameterTool.GetValue<int>(parameter, ElevenCode.UserLogin.ToString());
             Console.WriteLine(TmTimerTool.CurrentTime() + " TmEngineerMysql,userId:" + userId);
