+++ Servers/Mysqls/Components/Inventorys/TmKnapsackMysql.cs
@@ -7,7 +7,6 @@ namespace Servers
     {
         public override void TmAwake()
         {
-            base.TmAwake();
             DatabaseFormName = "knapsackitem";
         }
         public override void OnTransferParameter(object sender, TmParameter parameter)
@@ -19,27 +18,10 @@ namespace Servers
                     Console.WriteLine(TmTimerTool.CurrentTime() + " TmKnapsackMysql: " + elevenCode);
                     GetKnapsackByRolerId(sender, parameter);
                     break;
-                case (ElevenCode.Get):
-                    Console.WriteLine(TmTimerTool.CurrentTime() + " TmKnapsackMysql: " + elevenCode);
-                    GetInventory(sender, parameter);
-                    break;
                 case (ElevenCode.None):
                     break;
                 default:
                     break;
-            }
-        }
-        void GetInventory(object sender, TmParameter parameter)
-        {
-            Dictionary<int,TmInventory> inventorys = GetInventorys();
-            Console.WriteLine(TmTimerTool.CurrentTime() + " inventorys:" + inventorys.Count);
-            if (inventorys.Count > 0)
-            {
-                (sender as TmKnapsackHandler).Inventorys = inventorys;
-            }
-            else
-            {
-                Console.WriteLine(TmTimerTool.CurrentTime() + " 没有角色");
             }
         }
         void GetKnapsackByRolerId(object sender, TmParameter parameter)
