--- Servers/Mysqls/Components/Maps/Class1.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-namespace Servers.Mysqls.Components.Maps
-{
-    class Class1
-    {
-    }
-}
