+++ Servers/Mysqls/Components/Skills/TmBuffMysql.cs
@@ -11,7 +11,6 @@ namespace Servers
     {
         public override void TmAwake()
         {
-            base.TmAwake();
             DatabaseFormName = "buffitem";
         }
         public override void OnTransferParameter(object sender, TmParameter parameter)
@@ -23,27 +22,10 @@ namespace Servers
                     Console.WriteLine(TmTimerTool.CurrentTime() + " TmBuffMysql: " + elevenCode);
                     GetDbsByRolerId(sender, parameter);
                     break;
-                case (ElevenCode.Get):
-                    Console.WriteLine(TmTimerTool.CurrentTime() + " TmBuffMysql: " + elevenCode);
-                    GetSkill(sender, parameter);
-                    break;
                 case (ElevenCode.None):
                     break;
                 default:
                     break;
-            }
-        }
-        void GetSkill(object sender, TmParameter parameter)
-        {
-            Dictionary<int, TmSkill> skills = GetSkills();
-            Console.WriteLine(TmTimerTool.CurrentTime() + " skills:" + skills.Count);
-            if (skills!=null)
-            {
-                (sender as TmBuffHandler).Skills = skills;
-            }
-            else
-            {
-                Console.WriteLine(TmTimerTool.CurrentTime() + " 没有角色");
             }
         }
         void GetDbsByRolerId(object sender, TmParameter parameter)
