+++ Servers/Mysqls/Components/Skills/TmInbornMysql.cs
@@ -8,7 +8,6 @@ namespace Servers
     {
         public override void TmAwake()
         {
-            base.TmAwake();
             DatabaseFormName = "inbornitem";
         }
         public override void OnTransferParameter(object sender, TmParameter parameter)
